requirements.txt
# ============================================================================
# PROMETHEUS — Dependencies
# ============================================================================

# Core
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
PyYAML>=6.0

# Financial / Market Data
kiteconnect>=5.0.0
yfinance>=0.2.0
smartapi-python>=1.5.5

# Technical Analysis
ta>=0.10.0
pandas-ta>=0.3.14

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0

# AI — Local First (FREE)
ollama>=0.1.0
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0

# Optional: Google Gemini free tier
# google-generativeai>=0.3.0

# Backtesting
# backtrader>=1.9.78  # optional, we build custom engine

# Visualization
rich>=13.0.0
plotly>=5.15.0

# Telegram
python-telegram-bot>=20.0

# Database
sqlalchemy>=2.0.0

# Async
aiohttp>=3.8.0

# Scheduling
schedule>=1.2.0
APScheduler>=3.10.0

# Utilities
python-dateutil>=2.8.0
pyotp>=2.9.0
logzero>=1.7.0
websocket-client>=1.8.0
requests>=2.31.0
loguru>=0.7.0
tabulate>=0.9.0
